update svg demos
